+++ src/components/app/app.jsx
@@ -4,7 +4,8 @@ import AppHeader from "../app-header/app-header";
 import BurgerIngredients from "../burger-ingredients/burger-ingredients";
 import BurgerConstructor from "../burger-constructor/burger-constructor"
 import {useDispatch, useSelector} from "react-redux";
-import {ADD_BUN, ADD_INGREDIENT, getAllIngredients} from "../../services/actions";
+import {getAllIngredients} from "../../services/actions/all-ingredients";
+import {ADD_BUN, ADD_INGREDIENT} from "../../services/actions/burger-constructor";
 import {DndProvider} from "react-dnd";
 import {HTML5Backend} from "react-dnd-html5-backend";
 import { v4 as uuidv4 } from 'uuid';
@@ -12,13 +13,6 @@ import { v4 as uuidv4 } from 'uuid';
 function App() {
 
     const dispatch = useDispatch();
-
-    useEffect(
-        () => {
-            dispatch(getAllIngredients());
-        },
-        [dispatch]
-    );
 
     const onDropHandler = (item) => {
         const key = uuidv4();
