+++ src/services/actions/burger-constructor.jsx
@@ -1,0 +1,5 @@
+export const ADD_BUN = 'ADD_BUN';
+export const ADD_INGREDIENT = 'ADD_INGREDIENT';
+export const DEL_INGREDIENT = 'DEL_INGREDIENT';
+export const SORT_INGREDIENTS = 'SORT_INGREDIENTS';
+export const DROP_CONSTRUCTOR = 'DROP_CONSTRUCTOR';
