+++ src/services/actions/ingredient-details.jsx
@@ -1,0 +1,2 @@
+export const GET_INGREDIENT_DETAIL = 'GET_INGREDIENT_DETAIL';
+export const DEL_INGREDIENT_DETAIL = 'DEL_INGREDIENT_DETAIL';
