+++ src/services/reducers/all-ingredients.jsx
@@ -1,4 +1,4 @@
-import {GET_ALL_INGREDIENTS_FAILED, GET_ALL_INGREDIENTS_REQUEST, GET_ALL_INGREDIENTS_SUCCESS} from "../actions";
+import {GET_ALL_INGREDIENTS_FAILED, GET_ALL_INGREDIENTS_REQUEST, GET_ALL_INGREDIENTS_SUCCESS} from "../actions/all-ingredients";
 
 const initialData = {
     allIngredients: [],
@@ -24,7 +24,7 @@ export const allIngredientsReducer = (state = initialData, action) => {
         }
         case GET_ALL_INGREDIENTS_FAILED: {
             return {
-                ...state,
+                allIngredients: [],
                 ingredientsFailed: true,
                 ingredientsRequest: false
             };
